requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
apscheduler
httpx
python-dotenv
debugpy